refactor: moved otp into authentication
