+++ src/apps/backend/modules/authentication/internals/otp/store/otp-repository.ts
@@ -1,4 +1,4 @@
-import { ApplicationRepository } from '../../../application';
+import { ApplicationRepository } from '../../../../application';
 
 import { OTPDB, OTPDbSchema } from './otp-db';
