--- src/apps/backend/modules/otp/index.ts
@@ -1,2 +1,0 @@
-export { default as OTPService } from './otp-service';
-export * from './types';
